backend/pyproject.toml
[project]
name = "multi-tenant-ai-chatbot-platform-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "gunicorn>=21.2",
  "uvicorn[standard]>=0.27",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.1.0",
  "alembic>=1.13.0",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "pydantic[email]>=2.6",
  "pydantic-settings>=2.2",
  "openai>=1.0.0",
  "pgvector>=0.2.0",
  "tiktoken>=0.7.0",
]
[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
